Issue #OS-101 Validators module with Shex & json impl
