+++ java/registry/pom.xml
@@ -277,6 +277,11 @@
 			<version>1.0.0</version>
 		</dependency>
 		<dependency>
+			<groupId>io.opensaber</groupId>
+			<artifactId>validators</artifactId>
+			<version>1.0.0</version>
+		</dependency>
+		<dependency>
 			<groupId>commons-validator</groupId>
 			<artifactId>commons-validator</artifactId>
 			<version>1.6</version>
