--- java/registry/src/main/java/io/opensaber/registry/exception/RDFValidationException.java
@@ -1,11 +1,0 @@
-package io.opensaber.registry.exception;
-
-public class RDFValidationException extends Exception {
-
-    private static final long serialVersionUID = 3281835423990198453L;
-
-    public RDFValidationException(String message) {
-        super(message);
-    }
-
-}
