--- java/registry/src/main/java/io/opensaber/registry/exception/ValidationFactoryException.java
@@ -1,10 +1,0 @@
-package io.opensaber.registry.exception;
-
-public class ValidationFactoryException extends Exception {
-
-    private static final long serialVersionUID = -2123352045285196853L;
-
-    public ValidationFactoryException(String message) {
-        super(message);
-    }
-}
