--- java/registry/src/main/java/io/opensaber/registry/exception/errorconstants/ErrorConstants.java
@@ -1,14 +1,0 @@
-package io.opensaber.registry.exception.errorconstants;
-
-public class ErrorConstants {
-
-    //RDF validation error
-    public static final String RDF_VALIDATION_ERROR_MESSAGE = "Error while validating RDF";
-    public static final String RDF_DATA_IS_MISSING = "RDF Data is missing!";
-    public static final String RDF_DATA_IS_INVALID = "Data validation failed!";
-    public static final String SCHEMA_IS_NULL = "Schema for validation is missing";
-    public static final String VALIDATION_IS_MISSING = "Validation is missing";
-    public static final String INVALID_REQUEST_PATH = "Request URL is invalid";
-    public static final String VALIDATION_MISSING_FOR_TYPE = "Validation missing for type";
-    public static final String VALIDATION_IMPLEMENTATION_NOT_PROVIDED = "Validation implmentation not provided";
-}
