--- java/registry/src/main/java/io/opensaber/registry/service/ValidationService.java
@@ -1,12 +1,0 @@
-package io.opensaber.registry.service;
-
-import io.opensaber.pojos.ValidationResponse;
-import io.opensaber.registry.exception.RDFValidationException;
-import io.opensaber.registry.middleware.MiddlewareHaltException;
-
-import java.io.IOException;
-
-public interface ValidationService {
-
-    public ValidationResponse validateData(Object rdfModel, String origin) throws RDFValidationException, MiddlewareHaltException, IOException;
-}
