+++ java/schema-configuration/pom.xml
@@ -2,6 +2,26 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
+	<repositories>
+		<repository>
+			<snapshots>
+				<enabled>false</enabled>
+			</snapshots>
+			<id>bintray-labra-maven</id>
+			<name>bintray</name>
+			<url>https://dl.bintray.com/labra/maven</url>
+		</repository>
+	</repositories>
+	<pluginRepositories>
+		<pluginRepository>
+			<snapshots>
+				<enabled>false</enabled>
+			</snapshots>
+			<id>bintray-labra-maven</id>
+			<name>bintray-plugins</name>
+			<url>https://dl.bintray.com/labra/maven</url>
+		</pluginRepository>
+	</pluginRepositories>
 	<parent>
 		<groupId>io.opensaber</groupId>
 		<artifactId>opensaber</artifactId>
@@ -18,11 +38,19 @@
 	</properties>
 	
 	<dependencies>
+
 		<dependency>
-			<groupId>io.opensaber</groupId>
-			<artifactId>shaclex</artifactId>
-			<version>1.0.0</version>
-		</dependency>
+			<groupId>es.weso</groupId>
+			<artifactId>shaclex_2.12</artifactId>
+			<version>0.0.66</version>
+			<exclusions>
+				<exclusion>
+					<groupId>ch.qos.logback</groupId>
+					<artifactId>logback-classic</artifactId>
+				</exclusion>
+			</exclusions>
+			<type>pom</type>
+		</dependency> 
 		<dependency>
 			<groupId>io.opensaber</groupId>
 			<artifactId>middleware-commons</artifactId>
