+++ java/validators/pom.xml
@@ -1,20 +1,67 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <artifactId>opensaber</artifactId>
-        <groupId>io.opensaber</groupId>
-        <version>1.0.0</version>
-    </parent>
+<?xml version="1.0"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<repositories>
+		<repository>
+			<snapshots>
+				<enabled>false</enabled>
+			</snapshots>
+			<id>bintray-labra-maven</id>
+			<name>bintray</name>
+			<url>https://dl.bintray.com/labra/maven</url>
+		</repository>
+	</repositories>
+	<pluginRepositories>
+		<pluginRepository>
+			<snapshots>
+				<enabled>false</enabled>
+			</snapshots>
+			<id>bintray-labra-maven</id>
+			<name>bintray-plugins</name>
+			<url>https://dl.bintray.com/labra/maven</url>
+		</pluginRepository>
+	</pluginRepositories>
 
-    <artifactId>validators</artifactId>
-    <version>1.0.0</version>
-    <name>RDF Validation utilities</name>
-    <packaging>pom</packaging>
-    <modules>
-        <module>shex</module>
-    </modules>
+	<parent>
+		<groupId>io.opensaber</groupId>
+		<artifactId>opensaber</artifactId>
+		<version>1.0.0</version>
+	</parent>
+	<artifactId>validators</artifactId>
+	<version>1.0.0</version>
+	<packaging>jar</packaging>
+	<name>validators</name>
+	<url>http://maven.apache.org</url>
+
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+	</properties>
+
+	<dependencies>
+		<dependency>
+			<groupId>io.opensaber</groupId>
+			<artifactId>middleware-commons</artifactId>
+			<version>1.0.0</version>
+		</dependency>
+		<dependency>
+			<groupId>es.weso</groupId>
+			<artifactId>shaclex_2.12</artifactId>
+			<version>0.0.66</version>
+			<exclusions>
+				<exclusion>
+					<groupId>ch.qos.logback</groupId>
+					<artifactId>logback-classic</artifactId>
+				</exclusion>
+			</exclusions>
+			<type>pom</type>
+		</dependency>
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>3.8.1</version>
+			<scope>test</scope>
+		</dependency>
+	</dependencies>
 
 </project>
