--- java/validators/shex/pom.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <artifactId>validators</artifactId>
-    	<version>1.0.0</version>
-        <groupId>io.opensaber</groupId>
-    </parent>
-
-    <artifactId>shex</artifactId>
-    <version>1.0.0</version>
-    <name>ShEx based RDF Validation utilities</name>
-    <packaging>pom</packaging>
-    <modules>
-        <module>shaclex</module>
-    </modules>
-
-</project>
