--- java/validators/shex/shaclex/src/test/resources/good1.ttl
@@ -1,6 +1,0 @@
-prefix : <http://example.org/>
-@prefix sh: <http://www.w3.org/ns/shacl#>
-
-:x :p :y .
-
-:S sh:targetNode :x .
