--- java/validators/shex/shaclex/src/test/resources/teacher.jsonld
@@ -1,120 +1,0 @@
-{
-  "@context":{
-    "rdf":"http://www.w3.org/1999/02/22-rdf-syntax-ns#",
-    "rdfs":"http://www.w3.org/2000/01/rdf-schema#",
-    "teacher":"http://example.com/voc/teacher/1.0.0/",
-    "xsd":"http://www.w3.org/2001/XMLSchema#",
-    "schema": "http://schema.org/",
-    "@vocab":"http://example.com/voc/teacher/1.0.0/"
-  },
-  "@type":"Teacher",
-  "serialNum": 12,
-  "teacherCode":"12234",
-  "nationalIdentifier": {
-    "@value": 1234567890123456.1,
-    "@type": "xsd:decimal"
-  },
-  "teacherName":"Marvin Pande",
-  "gender":{
-    "@id":"teacher:GenderTypeCode-MALE"
-  },
-  "email": [{
-    "@type": "schema:Text",
-    "@value": "test1@gmail.com"
-  },
-  {
-    "@type": "schema:Text",
-    "@value": "test2@gmail.com"
-  }],
-  "birthDate": {
-    "@value": "1990-12-06",
-    "@type": "xsd:date"
-  },
-  "socialCategory":{
-    "@id":"teacher:SocialCategoryTypeCode-GENERAL"
-  },
-  "highestAcademicQualification":{
-    "@id":"teacher:AcademicQualificationTypeCode-PHD"
-  },
-  "highestTeacherQualification":{
-    "@id":"teacher:TeacherQualificationTypeCode-MED"
-  },
-  "yearOfJoiningService": {
-    "@value": "2014",
-    "@type" : "xsd:gYear"
-  },
-  "teacher:teachingRole":{
-      "teacher:mainSubjectsTaught":[
-        {
-          "@id":"teacher:SubjectCode-MATH"
-        },
-        {
-          "@id":"teacher:SubjectCode-PHYSICS"
-        }
-      ],
-      "teacher:teacherType":{
-        "@id":"teacher:TeacherTypeCode-HEAD"
-      },
-      "@type":"teacher:TeachingRole",
-      "teacher:appointedForSubjects":{
-        "@id":"teacher:SubjectCode-MATH"
-      },
-      "teacher:appointmentType":{
-        "@id":"teacher:TeacherAppointmentTypeCode-REGULAR"
-      },
-      "teacher:classesTaught":{
-        "@id":"teacher:ClassTypeCode-SECONDARYANDHIGHERSECONDARY"
-      },
-      "@id":"teacher:01255511-2ece-412d-8d43-010660331562",
-      "teacher:appointmentYear":{
-        "@value":"2015",
-        "@type":"xsd:gYear"
-      }
-  },
-  "basicProficiencyLevel":[
-      {
-        "@type":"BasicProficiencyLevel",
-        "proficiencySubject":{
-          "@id":"teacher:SubjectCode-SOCIALSTUDIES"
-        },
-        "proficiencyAcademicQualification":{
-          "@id":"teacher:AcademicQualificationTypeCode-BELOWSECONDARY"
-        }
-      }
-    ],
-  "inServiceTeacherTrainingFromBRC":{
-    "@type":"InServiceTeacherTrainingFromBlockResourceCentre",
-    "teacher:daysOfInServiceTeacherTraining": {
-      "@value": 10.8,
-      "@type": "xsd:decimal"
-    }
-  },
-  "inServiceTeacherTrainingFromCRC":{
-    "@type":"InServiceTeacherTrainingFromClusterResourceCentre",
-    "teacher:daysOfInServiceTeacherTraining": {
-      "@value": 2.3,
-      "@type": "xsd:decimal"
-    }
-  },
-  "inServiceTeacherTrainingFromDIET":{
-    "@type":"InServiceTeacherTrainingFromDIET",
-    "teacher:daysOfInServiceTeacherTraining": {
-      "@value": 5.7,
-      "@type": "xsd:decimal"
-    }
-  },
-  "inServiceTeacherTrainingFromOthers":{
-    "@type":"InServiceTeacherTrainingFromOthers",
-    "teacher:daysOfInServiceTeacherTraining": {
-      "@value": 3.5,
-      "@type": "xsd:decimal"
-    }
-  },
-  "nonTeachingAssignmentsForAcademicCalendar":{
-    "@type":"NonTeachingAssignmentsForAcademicCalendar",
-    "teacher:daysOfNonTeachingAssignments": {
-      "@value": 6,
-      "@type": "xsd:decimal"
-    }
-  }
-}
