+++ java/validators/src/main/java/io/opensaber/validators/core/ValidationService.java
@@ -1,0 +1,12 @@
+package io.opensaber.validators.core;
+
+import java.io.IOException;
+
+import io.opensaber.pojos.ValidationResponse;
+import io.opensaber.registry.middleware.MiddlewareHaltException;
+import io.opensaber.validators.exception.RDFValidationException;
+
+public interface ValidationService {
+
+    public ValidationResponse validateData(Object rdfModel, String origin) throws RDFValidationException,MiddlewareHaltException,IOException;
+}
