+++ java/validators/src/main/java/io/opensaber/validators/exception/RDFValidationException.java
@@ -1,0 +1,11 @@
+package io.opensaber.validators.exception;
+
+public class RDFValidationException extends Exception {
+
+    private static final long serialVersionUID = 3281835423990198453L;
+
+    public RDFValidationException(String message) {
+        super(message);
+    }
+
+}
