+++ java/validators/src/main/java/io/opensaber/validators/exception/ValidationFactoryException.java
@@ -1,0 +1,10 @@
+package io.opensaber.validators.exception;
+
+public class ValidationFactoryException extends Exception {
+
+    private static final long serialVersionUID = -2123352045285196853L;
+
+    public ValidationFactoryException(String message) {
+        super(message);
+    }
+}
